update examples/react-full/ to use data() instead of onBeforeRender()
