--- examples/react-full/pages/hello/+onBeforeRender.ts
@@ -1,19 +1,0 @@
-// https://vike.dev/onBeforeRender
-export { onBeforeRender }
-
-import type { OnBeforeRenderAsync } from 'vike/types'
-import { render } from 'vike/abort'
-import { names } from './names'
-
-const onBeforeRender: OnBeforeRenderAsync = async (pageContext): ReturnType<OnBeforeRenderAsync> => {
-  const { name } = pageContext.routeParams
-  if (name !== 'anonymous' && !names.includes(name)) {
-    throw render(404, `Unknown name: ${name}.`)
-  }
-  const pageProps = { name }
-  return {
-    pageContext: {
-      pageProps
-    }
-  }
-}
