+++ examples/react-full/pages/star-wars/@id/+dataIsomorph.ts
