+++ examples/react-full/pages/star-wars/index/+onBeforePrerenderStart.ts
@@ -12,13 +12,13 @@ const onBeforePrerenderStart: OnBeforePrerenderStartAsync = async (): ReturnType
     {
       url: '/star-wars',
       // We already provide `pageContext` here so that Vike
-      // will *not* have to call the `onBeforeRender()` hook defined
+      // will *not* have to call the `data()` hook defined
       // above in this file.
       pageContext: {
-        pageProps: {
-          movies: filterMoviesData(movies)
-        },
-        title: getTitle(movies)
+        data: {
+          movies: filterMoviesData(movies),
+          title: getTitle(movies)
+        }
       }
     },
     ...movies.map((movie) => {
@@ -27,13 +27,13 @@ const onBeforePrerenderStart: OnBeforePrerenderStartAsync = async (): ReturnType
         url,
         // Note that we can also provide the `pageContext` of other pages.
         // This means that Vike will not call any
-        // `onBeforeRender()` hook and the Star Wars API will be called
+        // `data()` hook and the Star Wars API will be called
         // only once (in this `prerender()` hook).
         pageContext: {
-          pageProps: {
-            movie: filterMovieData(movie)
-          },
-          title: movie.title
+          data: {
+            movie: filterMovieData(movie),
+            title: movie.title
+          }
         }
       }
     })
