+++ examples/react-full/renderer/+config.h.ts
@@ -2,7 +2,7 @@ import type { Config } from 'vike/types'
 
 // https://vike.dev/config
 export default {
-  passToClient: ['pageProps', 'title', 'someAsyncProps'],
+  passToClient: ['title', 'someAsyncProps'],
   clientRouting: true,
   hydrationCanBeAborted: true,
   // https://vike.dev/meta
@@ -11,8 +11,8 @@ export default {
     title: {
       env: { server: true, client: true }
     },
-    // Create new config 'onBeforeRenderIsomorph'
-    onBeforeRenderIsomorph: {
+    // Create new config 'dataIsomorph'
+    dataIsomorph: {
       env: { config: true },
       effect({ configDefinedAt, configValue }) {
         if (typeof configValue !== 'boolean') {
@@ -21,9 +21,9 @@ export default {
         if (configValue) {
           return {
             meta: {
-              onBeforeRender: {
-                // We override Vike's default behavior of always loading/executing onBeforeRender() on the server-side.
-                // If we set onBeforeRenderIsomorph to true, then onBeforeRender() is loaded/executed in the browser as well, allowing us to fetch data direcly from the browser upon client-side navigation (without involving our Node.js/Edge server at all).
+              data: {
+                // We override Vike's default behavior of always loading/executing data() on the server-side.
+                // If we set dataIsomorph to true, then data() is loaded/executed in the browser as well, allowing us to fetch data direcly from the browser upon client-side navigation (without involving our Node.js/Edge server at all).
                 env: { server: true, client: true }
               }
             }
@@ -33,7 +33,7 @@ export default {
     }
   },
   hooksTimeout: {
-    onBeforeRender: {
+    data: {
       error: 30 * 1000,
       warning: 10 * 1000
     }
